feat: Added Tests 🧪 & Removed Duplicate Models
