+++ pydantic_models/Adp/native_sad.py
@@ -1,0 +1,78 @@
+# generated by datamodel-codegen:
+#   filename:  native-sad-1-of-3
+#   timestamp: 2023-11-19T08:30:04+00:00
+
+from __future__ import annotations
+from re import A
+
+from typing import Any, List, Union
+
+from pydantic import BaseModel, Field
+from pydantic_models.Adp.adpexecutionstack import AdpExecutionStack
+from pydantic_models.Adp.fastxds import (Gopy, StatusHistoryBase, FQDNBase, AngularRangeBase, AdpStatusBase, AdpInfoBase)
+from pydantic_models.Adp.xdsinputparams import (Result, FastXds2Item, FastXds3Item, FastXds3Params, GopyItem, GopyParams)
+
+class StatusHistory(StatusHistoryBase):
+    gopy: Gopy
+
+
+class FQDN(FQDNBase):
+    gopy: str
+
+
+class AngularRange(AngularRangeBase):
+    gopy: float
+
+
+class AdpStatus(AdpStatusBase):
+    gopy: str
+
+
+class AdpInfo(AdpInfoBase):
+    gopy: str
+
+
+class Metadatum(BaseModel):
+    name: str
+    value: Union[List[Any], float, str]
+
+
+class Model(BaseModel):
+    field_id: str = Field(..., alias='_id')
+    userAccount: str
+    user: str
+    uid: int
+    gid: int
+    createdOn: str
+    lastUpdated: str
+    beamline: str
+    trackingId: str
+    method: str
+    beamTransmission: float
+    beam_position_x: float
+    beam_position_y: float
+    detectorDistance: int
+    exposureTime: float
+    folder: str
+    root_folder: str
+    numFramesTotal: int
+    datasetNo: int
+    numberOfExpectedDatasets: int
+    oscillationAngle: float
+    prefix: str
+    runNumber: int
+    wavelength: float
+    adpExecutionStack: AdpExecutionStack
+    statusHistory: StatusHistory
+    FQDN: FQDN
+    angularRange: AngularRange
+    adpStatus: AdpStatus
+    adpInfo: AdpInfo
+    result: Result
+    detector: str
+    roi: str
+    mergeFolder: str
+    mergeId: str
+    seen: bool
+    datasets_id: None
+    metadata: List[Metadatum]
