+++ pydantic_models/Adp/standard_xia2dials.py
@@ -1,0 +1,127 @@
+# generated by datamodel-codegen:
+#   filename:  standard
+#   timestamp: 2023-11-19T08:09:53+00:00
+
+from __future__ import annotations
+from re import A
+
+from typing import Any, List, Union
+
+from pydantic import BaseModel, Field
+from pydantic_models.Adp.adpexecutionstack import AdpExecutionStack
+from pydantic_models.Adp.fastxds import (Xia2dials, StatusHistoryBase, FQDNBase, AngularRangeBase, AdpStatusBase, AdpInfoBase)
+from pydantic_models.Adp.xdsinputparams import FastXds2Item, FastXds3Item, FastXds3Params
+
+
+class StatusHistory(StatusHistoryBase):
+    xia2dials: Xia2dials
+
+
+class FQDN(FQDNBase):
+    xia2dials: str
+
+
+class AngularRange(AngularRangeBase):
+    xia2dials: float
+
+
+class AdpStatus(AdpStatusBase):
+    xia2dials: str
+
+
+class AdpInfo(AdpInfoBase):
+    xia2dials: str
+
+
+
+
+class Xia2dial(BaseModel):
+    order: str
+    low_resolution_limit: float
+    resolution_limit: float
+    r_factor_observed: float
+    rmeas_within: float
+    rmeas: float
+    rpim_within: float
+    rpim: float
+    observed_reflections: int
+    unique_reflections: int
+    Isigma: float
+    completeness: float
+    multiplicity: float
+    cc_half: float
+    anomalous_completeness: float
+    anomalous_multiplicity: float
+    anomalous_correlation: float
+
+
+class Xia2dialsParams(BaseModel):
+    command: str
+    spaceGroupNumber: int
+    spaceGroupLabel: str
+    a: str
+    b: str
+    c: str
+    alpha: str
+    beta: str
+    gamma: str
+    mosaicity: float
+    isa: str
+    wilsonBfactor: float
+
+
+class Result(BaseModel):
+    fast_xds_2: List[FastXds2Item]
+    fast_xds_3: List[FastXds3Item]
+    fast_xds_3_params: FastXds3Params
+    xia2dials: List[Xia2dial]
+    xia2dials_params: Xia2dialsParams
+    fast_xds_1: List
+    fast_xds_1_params: List
+    fast_xds_2_params: List
+
+
+class Metadatum(BaseModel):
+    name: str
+    value: Union[List[Any], float, str]
+
+
+class Model(BaseModel):
+    field_id: str = Field(..., alias='_id')
+    userAccount: str
+    user: str
+    uid: int
+    gid: int
+    createdOn: str
+    lastUpdated: str
+    beamline: str
+    trackingId: str
+    method: str
+    beamTransmission: float
+    beam_position_x: float
+    beam_position_y: float
+    detectorDistance: float
+    exposureTime: float
+    folder: str
+    root_folder: str
+    numFramesTotal: int
+    datasetNo: int
+    numberOfExpectedDatasets: int
+    oscillationAngle: float
+    prefix: str
+    runNumber: int
+    wavelength: float
+    adpExecutionStack: AdpExecutionStack
+    statusHistory: StatusHistory
+    FQDN: FQDN
+    angularRange: AngularRange
+    adpStatus: AdpStatus
+    adpInfo: AdpInfo
+    result: Result
+    detector: str
+    roi: str
+    mergeFolder: str
+    mergeId: str
+    seen: bool
+    datasets_id: None
+    metadata: List[Metadatum]
